Move activities into a specific package
