+++ app/src/main/java/com/example/checnutritionapp/ui/user_profile/UserProfileFragment.java
@@ -1,7 +1,6 @@
 package com.example.checnutritionapp.ui.user_profile;
 
 import android.os.Bundle;
-import android.util.StringBuilderPrinter;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -11,13 +10,11 @@ import android.widget.TextView;
 import androidx.annotation.Nullable;
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
-import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
 
-import com.example.checnutritionapp.MainActivity;
+import com.example.checnutritionapp.activities.MainActivity;
 import com.example.checnutritionapp.R;
 import com.example.checnutritionapp.model.User;
-import com.example.checnutritionapp.utility.UserSet;
 
 public class UserProfileFragment extends Fragment {
