+++ app/src/main/res/layout/place_order_activity.xml
@@ -5,4 +5,4 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PlaceOrderActivity"/>
+    tools:context=".activities.PlaceOrderActivity"/>
