+++ app/src/main/res/layout/summary_activity.xml
@@ -4,4 +4,4 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SummaryActivity" />
+    tools:context=".activities.SummaryActivity" />
